Add files via upload
dataset.py for hyper that has theta_p as an edge attribute. Requires parsing the data file so that nodes have rapidity, pull_rapidity and pull_phi as attributes in order for the theta_p to be calculated. This parser is also in this file. However, parser doesnt currently work as adam did a terrible job of it.